Update index.html
Prompt User for Number of Players:

When the page loads, prompt the user to input how many players are playing the game.
The user should still be able to add more players later using the "Add Player" button.
Prompt for Default Buy-In Value:

After asking for the number of players, prompt the user whether they want to set a default buy-in value for all players.
If the user provides a value, it will be applied to all cells in the second column (Buy-In). If the user chooses to enter buy-ins manually, they can edit each player's buy-in as needed.
Footnote Update:

Add the footnote "v0.12" at the bottom of the page and make it visible.