Update index.html
Key Changes:
Sequential Name Input:

After the user specifies the number of players, they are prompted for each player's name in sequence. The names are added directly to the table as the user progresses through the prompts.
If the user does not enter a name, a default name like Player 1, Player 2, etc., will be assigned.
Version Number:

The version number at the bottom has been updated to v0.1.3.
How It Works:
Page Load:

Upon loading, the user is first asked how many players are playing, then sequentially asked for each player's name.
Default Buy-In:

If the user chooses to set a default buy-in, the buy-in value will be applied to all players. If not, the user can manually input the buy-in values for each player.
Footnote:

The footnote with the version number is clearly visible at the bottom of the page.